Example layout and certificate layout change
Example layout in the layout editing page must follow the participation
or appreciation items because it's taller than participation layout.
This is to make certificate layout is more related to how editor layout
looks.

For certificate layout, if there's no category, DON'T show it.
Not even the defaultText AKA testing/example text.

Close #92